# api/openapi.yaml
openapi: 3.0.3
info:
  title: GenomeVault API
  version: 1.0.0
  description: |
    Privacy-preserving genomic computing platform using hyperdimensional computing,
    zero-knowledge proofs, and federated learning for secure genomic data analysis.

    ## Privacy Model

    GenomeVault uses mathematical privacy guarantees rather than traditional encryption:
    - **Hyperdimensional Computing**: Genomic variants encoded as high-dimensional vectors
    - **Private Information Retrieval (PIR)**: Query databases without revealing what you're looking for
    - **Zero-Knowledge Proofs**: Verify computations without revealing the data
    - **Differential Privacy**: Mathematical bounds on information leakage

    ## Rate Limits

    API requests are subject to rate limiting:
    - **Standard tier**: 1000 requests/hour per API key
    - **Clinical tier**: 10000 requests/hour per API key
    - **Research tier**: 50000 requests/hour per API key

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Request limit per window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Time when rate limit resets (Unix timestamp)

  contact:
    name: GenomeVault Support
    url: https://github.com/genomevault/genomevault
    email: support@genomevault.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  x-logo:
    url: "https://genomevault.io/logo.png"
    altText: "GenomeVault Logo"
servers:
...
paths:
...
  /hv/encode:
    post:
      tags:
        - Hypervector
      summary: Encode genomic data to hypervector
      description: |
        Convert genomic variants or numeric features into high-dimensional vectors
        for privacy-preserving computation.

        ## Input Types
        - **Variants**: VCF-like variant objects with chromosome, position, ref, alt
        - **Numeric**: Pre-processed numeric feature arrays

        ## Privacy Guarantees
        - Input data is not stored on servers
        - Hypervectors provide k-anonymity through high-dimensional encoding
        - Original genomic data cannot be reconstructed from hypervectors
      operationId: encodeHypervector
      security:
        - ApiKeyAuth: []
        - OAuth2: [genomic:read]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncodeRequest'
            examples:
              variants_example:
                summary: Encode genomic variants
                value:
                  variants:
                    - chrom: "1"
                      pos: 1234567
                      ref: "A"
                      alt: "T"
                      impact: "missense"
                    - chrom: "2"
                      pos: 9876543
                      ref: "G"
                      alt: "C"
                      impact: "synonymous"
                  dim: 8192
                  binary: false
              numeric_example:
                summary: Encode numeric features
                value:
                  numeric: [0.1, 0.8, 0.3, 0.9, 0.2]
                  dim: 4096
                  binary: true
      responses:
        '200':
          description: Successfully encoded data
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncodeResponse'
              example:
                dim: 8192
                binary: false
                vector: [0.12, -0.34, 0.56, 0.78]
                privacy_level: "k-anonymous"
                compression_ratio: 87.3
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /pir/query:
    post:
      tags:
        - Private Information Retrieval
      summary: Execute PIR query
      description: |
        Perform private information retrieval query without revealing the query index
        to the server. This maintains query privacy while retrieving genomic data.

        ## Privacy Guarantees
        - Server cannot determine which record was queried
        - Cryptographic privacy protection using lattice-based PIR
        - Query patterns are not logged or tracked
      operationId: pirQuery
      security:
        - ApiKeyAuth: []
        - OAuth2: [pir:query]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PIRQueryRequest'
            example:
              index: 42
              query_id: "unique-query-identifier"
      responses:
        '200':
          description: PIR query executed successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PIRQueryResponse'
              example:
                index: 42
                item_base64: "YWxwaGE="
                privacy_proof: "zk_proof_hash_example"
                query_time_ms: 125
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for authentication. Get your API key from the GenomeVault console.

        Example: `X-API-Key: gv_1234567890abcdef`

    OAuth2:
      type: oauth2
      description: |
        OAuth 2.0 with PKCE for secure authentication. Supports the following scopes:

        - `genomic:read` - Read genomic data and perform basic encoding
        - `pir:query` - Execute private information retrieval queries
        - `zk:prove` - Generate and verify zero-knowledge proofs
        - `clinical:analyze` - Perform clinical genomic analysis
        - `admin:manage` - Administrative operations

      flows:
        authorizationCode:
          authorizationUrl: https://auth.genomevault.io/oauth/authorize
          tokenUrl: https://auth.genomevault.io/oauth/token
          scopes:
            genomic:read: Read genomic data
            pir:query: Execute PIR queries
            zk:prove: Generate ZK proofs
            clinical:analyze: Clinical analysis
            admin:manage: Administrative access
  headers:
    X-RateLimit-Limit:
      description: Request limit per time window
      schema:
        type: integer
        example: 1000
    X-RateLimit-Remaining:
      description: Requests remaining in current window
      schema:
        type: integer
        example: 999
    X-RateLimit-Reset:
      description: Time when rate limit resets (Unix timestamp)
      schema:
        type: integer
        example: 1642248600
  schemas:
...
    EncodeRequest:
      type: object
      properties:
        numeric:
          type: array
          items:
            type: number
          description: Numeric feature array (alternative to variants)
          example: [0.1, 0.8, 0.3, 0.9]
        variants:
          type: array
          items:
            $ref: '#/components/schemas/GenomicVariant'
          description: Genomic variants to encode (alternative to numeric)
        dim:
          type: integer
          default: 8192
          minimum: 1024
          maximum: 100000
          description: Hypervector dimension
        binary:
          type: boolean
          default: false
          description: Return binary (-1/+1) or continuous values
      oneOf:
        - required: [numeric]
        - required: [variants]
    GenomicVariant:
      type: object
      required: [chrom, pos, ref, alt]
      properties:
        chrom:
          type: string
          description: Chromosome (1-22, X, Y, M)
          pattern: "^(chr)?(1[0-9]|2[0-2]|[1-9]|X|Y|M|MT)$"
          example: "1"
        pos:
          type: integer
          minimum: 1
          description: Genomic position (1-based)
          example: 1234567
        ref:
          type: string
          pattern: "^[ATCGN]+$"
          description: Reference allele
          example: "A"
        alt:
          type: string
          pattern: "^[ATCGN]+$"
          description: Alternative allele
          example: "T"
        impact:
          type: string
          enum: [missense, nonsense, synonymous, frameshift, splice_site, intron, intergenic]
          description: Predicted functional impact
          example: "missense"
        quality:
          type: number
          minimum: 0
          maximum: 100
          description: Variant quality score
          example: 99.5
    EncodeResponse:
      type: object
      required: [dim, binary, vector]
      properties:
        dim:
          type: integer
          description: Hypervector dimension
        binary:
          type: boolean
          description: Whether vector contains binary values
        vector:
          type: array
          items:
            oneOf:
              - type: integer
              - type: number
          description: Encoded hypervector
        privacy_level:
          type: string
          enum: [k-anonymous, differential_private, information_theoretic]
          description: Privacy guarantee level
        compression_ratio:
          type: number
          description: Data compression ratio achieved
          example: 87.3
    PIRQueryRequest:
      type: object
      required: [index]
      properties:
        index:
          type: integer
          minimum: 0
          description: Index to query (kept private from server)
        query_id:
          type: string
          format: uuid
          description: Unique query identifier for tracking
        timeout_seconds:
          type: integer
          default: 30
          minimum: 1
          maximum: 300
          description: Query timeout
    PIRQueryResponse:
      type: object
      required: [index, item_base64]
      properties:
        index:
          type: integer
          description: Queried index (for client verification)
        item_base64:
          type: string
          format: byte
          description: Base64-encoded retrieved item
        privacy_proof:
          type: string
          description: Cryptographic proof of privacy preservation
        query_time_ms:
          type: integer
          description: Query execution time in milliseconds
...
    Error:
      type: object
      required: [type, code, message]
      properties:
        type:
          type: string
          description: Error type classification
          example: "ValidationError"
        code:
          type: string
          description: Machine-readable error code
          example: "GV_INVALID_INPUT"
        message:
          type: string
          description: Human-readable error message (PHI-safe)
          example: "Invalid genomic coordinate format"
        details:
          type: object
          description: Additional error context
          properties:
            request_id:
              type: string
              format: uuid
              description: Unique request identifier for support
            field:
              type: string
              description: Field that caused validation error
            allowed_values:
              type: array
              items:
                type: string
              description: List of allowed values
        trace_id:
          type: string
          description: Distributed tracing identifier
  responses:
    BadRequest:
      description: Invalid request parameters
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "ValidationError"
            code: "GV_INVALID_INPUT"
            message: "Invalid genomic coordinate format"
            details:
              request_id: "req_1234567890"
              field: "variants[0].chrom"
              allowed_values: ["1", "2", "3", "...", "22", "X", "Y", "M"]
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "AuthenticationError"
            code: "GV_UNAUTHORIZED"
            message: "Valid authentication required"
            details:
              request_id: "req_1234567890"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          description: Seconds until rate limit resets
          schema:
            type: integer
            example: 3600
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "RateLimitError"
            code: "GV_RATE_LIMITED"
            message: "Rate limit exceeded"
            details:
              request_id: "req_1234567890"
              limit: 1000
              window_seconds: 3600
              retry_after: 3600
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            type: "InternalServerError"
            code: "GV_INTERNAL_ERROR"
            message: "An unexpected error occurred"
            details:
              request_id: "req_1234567890"
...